# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Configuration for a single instrument to score
 */
export interface InstrumentConfig {
  /** Short instrument name shown in the UI (e.g., BTCUSD) */
  name: string;
  /** Display ticker label (e.g., BTC-USD) */
  displayTicker: string;
  /** Polygon.io ticker symbol (e.g., X:BTCUSD) */
  polyTicker: string;
}

/**
 * Color category based on score
 */
export type InstrumentScoreColorCategory =
  (typeof InstrumentScoreColorCategory)[keyof typeof InstrumentScoreColorCategory];

export const InstrumentScoreColorCategory = {
  green: "green",
  yellow: "yellow",
  red: "red",
} as const;

export interface InstrumentScore {
  /** Instrument display name */
  name: string;
  /** Display ticker symbol */
  ticker: string;
  /** Final display score 0-100 (tie-adjusted) */
  score: number;
  /** Base environment score before tie-breaking */
  rawScore: number;
  /** Tie-breaker sub-score 0-10 */
  subScore: number;
  /** Average True Range value */
  atr: number;
  /** Average Directional Index value */
  adx: number;
  /** 7-day percentage return */
  momentum: number;
  /** Average hourly volume over 30 days */
  avgVolume: number;
  /** Short explanation of the score including tie-break reason if applicable */
  explanation: string;
  /** Risk management guidance based on score */
  riskGuidance: string;
  /** Color category based on score */
  colorCategory: InstrumentScoreColorCategory;
}

export interface MarketScoresResponse {
  instruments: InstrumentScore[];
  /** Name of the highest scoring instrument */
  bestInstrument: string;
  /** Score of the best instrument */
  bestScore: number;
  /** ISO timestamp of when data was fetched */
  lastUpdated: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetMarketScoresParams = {
  /**
 * JSON-encoded array of instrument configs to score. Each item must have `name`, `displayTicker`, and `polyTicker`. If omitted, defaults to BTCUSD / XAUUSD / EURUSD.

 */
  instruments?: string;
};