# lib/api-zod/src/generated/types/instrumentConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Configuration for a single instrument to score
 */
export interface InstrumentConfig {
  /** Short instrument name shown in the UI (e.g., BTCUSD) */
  name: string;
  /** Display ticker label (e.g., BTC-USD) */
  displayTicker: string;
  /** Polygon.io ticker symbol (e.g., X:BTCUSD) */
  polyTicker: string;
}