# lib/api-zod/src/generated/types/instrumentScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InstrumentScoreColorCategory } from "./instrumentScoreColorCategory";

export interface InstrumentScore {
  /** Instrument display name */
  name: string;
  /** Display ticker symbol */
  ticker: string;
  /** Final display score 0-100 (tie-adjusted) */
  score: number;
  /** Base environment score before tie-breaking */
  rawScore: number;
  /** Tie-breaker sub-score 0-10 */
  subScore: number;
  /** Average True Range value */
  atr: number;
  /** Average Directional Index value */
  adx: number;
  /** 7-day percentage return */
  momentum: number;
  /** Average hourly volume over 30 days */
  avgVolume: number;
  /** Short explanation of the score including tie-break reason if applicable */
  explanation: string;
  /** Risk management guidance based on score */
  riskGuidance: string;
  /** Color category based on score */
  colorCategory: InstrumentScoreColorCategory;
}