# lib/api-zod/src/generated/types/marketScoresResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InstrumentScore } from "./instrumentScore";

export interface MarketScoresResponse {
  instruments: InstrumentScore[];
  /** Name of the highest scoring instrument */
  bestInstrument: string;
  /** Score of the best instrument */
  bestScore: number;
  /** ISO timestamp of when data was fetched */
  lastUpdated: string;
}